Fix Grappa::init() global-heap size, fixes GlobalAllocator test

+++ system/Grappa.hpp
@@ -43,7 +43,7 @@ namespace Grappa {
   
   /// Initialize Grappa. Call in SPMD context before running Grappa
   /// code. Running Grappa code before calling init() is illegal.
-  void init( int * argc_p, char ** argv_p[], size_t size = -1 );
+  void init( int * argc_p, char ** argv_p[], int64_t size = -1 );
 
   /// Clean up Grappa. Call in SPMD context after all Grappa code
   /// finishes. Running Grappa code after calling finalize() is illegal.
